Change how alternative symbol definitions are stored
We previously assumed that there wouldn't be many symbols with multiple
definitions, so stored alternatives in a HashMap<SymbolId,
Vec<SymbolId>> for any symbol with more than one defintions. However
when linking C++ code, there can be heaps of symbols with multiple weak
definitions.

Instead, we now store effectively a link list of symbol IDs that have
the same name. This allows us to store all alternatives in a single vec.
Plus one more vec that stores the roots of all chains.