Initialise symbol resolutions vec from separate threads
Previously, the main thread initialised the vec, filling it with all
None, then the separate threads got mutable slices of the vec and
replace the None values.

Now, the vec is allocated but not initialised by the main thread and the
initialisation happens as the resolutions are computed by the separate
threads.