Use a hashmap for recording custom sections rather than a BTreeMap
It's easier to get good performance with a hashmap. It's not clear why I
originally used a BTreeMap. It was possibly related to determinism, but
we can still get deterministic behaviour with a hashmap, so long as we
allocate IDs based on the order in which we first see a section name and
not on iteration order of the map. So that's what we now do.

Made the code slightly more readable while I was at it.