Large refactoring: Introduce PartId
Prior to this change, we didn't have an ID to represent a part of a
section. OutputSectionPartMap stored each part that needed to separately
identified in a separate field. This meant that adding a new section
required changes in lots of places.

There's still more that needs to be done, but this should be the largest
change required for issue #36.