Split section resolution from symbol resolution
If this hurts performance, it's only by a tiny amount, such that it's not
easily measurable. Splitting it means that we can see the time to resolve
sections separately from symbol resolution. The main advantage however is that
it opens the option for storing more data on groups rather than objects. Once
symbol resolution is complete, we know which objects we're loading, which means
in section resolution, we can store data onto the group without getting
non-deterministic ordering.