Delete old output file after we've created the new one
This avoids the need to use a separate thread. The extra thread was undesirable
because if the user specifies `--threads=N`, then we should limit ourselves to
N active threads at any one time and with deletion on its own thread we had
`N+1` active threads for a time.

When running single-threaded, we now leave deleting the old output file until
just before we're going to create the new one.