Move forking code into its own module in wild_lib
Forking uses unsafe. By putting all forking-related code together in its own
module, it's clearer where the boundary of the related unsafety is.

Additionally, this should make it easier to make forking an optional feature,
which might help with porting to platforms that don't have fork.

This could have been a module of wild (the binary crate) rather than a module
of wild_lib, however by moving it to the former, the done_callback mechanism
can be made private to wild_lib.