Change the way we test and verify allocation consistency
The motivation for this change is that `are_flags_valid` is hard to get right
and hard to keep up-to-date. The plan is to delete it and the test that depends
on it - `test_resolution_allocation_consistency`. However, when inconsistent
allocations occur, it can be hard to figure out what combination of flags
caused the inconsistent allocation. So we reuse the consistency checking
mecanism used by the test from the linker proper, but only when an environment
variable is set. When we detect an inconsistency at write-time, we now suggest
setting this environment variable to get more info.