Use i64 rather than u64 for addend
This makes it less likely to get mixed up with offset which is often an
adjacent argument. It also makes it nicer if we need to print it out for
debugging purposes.

We still cast to u64 before calling `.wrapping_add` when we apply the addend,
but now we just do it at the point where we're applying the addend.