Only use GLOB_DAT for GOT entries
The ELF spec says that GLOB_DAT resolves to the address of the symbol and
doesn't mention applying any addend. Some runtimes do apply an addend, but we
can't rely on that. So only use GLOB_DAT for GOT entries, which don't have an
addend. For other absolute symbol relocations, use R_x86_64_64.

Fixes #576