Refactoring: Move keep property to section rules
This effectively makes it a property of the input section rather than the
output section, which is more accurate to how it's supposed to work. e.g. a
linker script could in theory specify to keep some sections and not others,
while putting all of them into the same output section.

Issue #44