save-dir: Replicate original directory structure
This makes it easier to see where files came from originally. It also makes it
possible for things like sysroot, which relies on directory structure to work
as expected.

We also now copy input files that aren't named by path on the command line.
e.g. libraries that are named via -l.

Where files are obtained via symlinks, we copy those symlinks in addition to
the file that they point to.

Fixes #575