Merge ResolutionFlags and ValueFlags (#1180)
The two were previously stored separately, but we often need them at the same
time. Previously they each were stored as a u8, now together they're stored as
a u16. So it takes the same amount of space, but we have more flexibilty about
what we use free bits for. In particular, ResolutionFlags was previously full,
so we couldn't add any additional resolution flags. Now we can.

Perhaps more importantly, we now have the option of packing bits of what was
previously ResolutionFlags based on a bit in what was previously ValueFlags.
Specifically, we could overlap GOT,PLT with
GOT_TLS_MODULE,GOT_TLS_OFFSET,GOT_TLS_DESCRIPTOR and select between the two
based on a new bit that indicates if the value is a TLS.

The main downside of this change is that it expands the scope during which
SymbolDb doesn't have direct access to what was ValueFlags. There's a TODO on
take_flags_atomic to try to fix this.