perf: Rewrite scheduling of string merging tasks (#1240)
The old do_splitting_work, if there were lots of threads and insufficient work
to do, could end up effectively busy-waiting, which was terrible for
performance.

We now spawn a separate task for each bit of work. We also now constrain the
number of split input section Vecs that we'll keep in memory.

Issue #1085