feat: Support `-z max-page-size` (#1400)
This option is used when building
[gVisor](https://github.com/google/gvisor/blob/89151a08e48f15c73ae24d5d460f1d6400c86675/vdso/BUILD#L49).

This is a prefix-style option, but the existing argument parser did not
support prefix options that take a value. To address this, the second
commit refactors the option handling logic to distinguish between
sub-options that take a value and those that do not.

In addition, for testing, this also introduces a new test directive,
`ExpectLoadAlignment`, which checks the alignment of the first `PT_LOAD`
segment in the test binary.